Use org-cliplink to paste URLs.
Actually, define a custom function based on org-cliplink that tune the
title used for GitLab/GitHub issues/commits/pull-requests/merge-requests
to follow the same format as we were using before.

This allows to copy the URL to the clipboard and paste here through this
fucntion to get a nicely formatted link.